Fix the problem of multi-apps in Firebase.
