+++ src/components/pages/Selling.vue
@@ -170,7 +170,7 @@
   import ElInput from "../../../node_modules/element-ui/packages/input/index";
   import ElButton from "../../../node_modules/element-ui/packages/button/src/button";
   import ElCard from "../../../node_modules/element-ui/packages/card/src/main";
-  import firebase from "firebase";
+  // import firebase from "firebase";
 
   //  let config = {
   //    apiKey: 'AIzaSyDxF-ZmWZP0qtgeQQli2oPCTn6hw4aovbo',
@@ -181,17 +181,17 @@
   //    messagingSenderId: '193615524398'
   //  }
 
-  let config = {
-    apiKey: "AIzaSyBhr-1yhcgKdqiyYwPRWJ2DatlzxzonHhs",
-    authDomain: "se-retail-management.firebaseapp.com",
-    databaseURL: "https://se-retail-management.firebaseio.com",
-    projectId: "se-retail-management",
-    storageBucket: "se-retail-management.appspot.com",
-    messagingSenderId: "729149844491"
-  };
+  // let config = {
+  //   apiKey: "AIzaSyBhr-1yhcgKdqiyYwPRWJ2DatlzxzonHhs",
+  //   authDomain: "se-retail-management.firebaseapp.com",
+  //   databaseURL: "https://se-retail-management.firebaseio.com",
+  //   projectId: "se-retail-management",
+  //   storageBucket: "se-retail-management.appspot.com",
+  //   messagingSenderId: "729149844491"
+  // };
 
-  let app = firebase.initializeApp(config);
-  let db = app.database();
+  // let app = firebase.initializeApp(config);
+  import db from '../common/Firebase'
   let dbRef = {
     source: db.ref("categories")
     // asObject: true
