+++ src/main.js
@@ -7,6 +7,7 @@ import Firebase from 'firebase'
 import router from './router'
 import ElementUI from 'element-ui'
 import 'element-ui/lib/theme-default/index.css'
+// import 'element-ui/lib/theme-chalk/index.css'
 
 Vue.use(ElementUI)
 Vue.use(VueFire)
